requirements.txt
Flask==2.1.1
openai==0.11.0
pinecone-client==2.1.5

